fix: #32; refactor sortList() moving to util
